+++ examples/playground/prompt.py
@@ -9,6 +9,7 @@ from llm_guard.input_scanners import (
     BanSubstrings,
     BanTopics,
     Code,
+    Language,
     PromptInjection,
     PromptInjectionV2,
     Regex,
@@ -29,6 +30,7 @@ def init_settings() -> (List, Dict):
         "BanSubstrings",
         "BanTopics",
         "Code",
+        "Language",
         "PromptInjection",
         "PromptInjectionV2",
         "Regex",
@@ -171,6 +173,79 @@ def init_settings() -> (List, Dict):
         settings["Code"] = {
             "languages": st_cd_languages,
             "mode": st_cd_mode,
+        }
+
+    if "Language" in st_enabled_scanners:
+        st_lan_expander = st.sidebar.expander(
+            "Language",
+            expanded=False,
+        )
+
+        with st_lan_expander:
+            st_lan_valid_language = st.multiselect(
+                "Languages",
+                [
+                    "af",
+                    "ar",
+                    "bg",
+                    "bn",
+                    "ca",
+                    "cs",
+                    "cy",
+                    "da",
+                    "de",
+                    "el",
+                    "en",
+                    "es",
+                    "et",
+                    "fa",
+                    "fi",
+                    "fr",
+                    "gu",
+                    "he",
+                    "hi",
+                    "hr",
+                    "hu",
+                    "id",
+                    "it",
+                    "ja",
+                    "kn",
+                    "ko",
+                    "lt",
+                    "lv",
+                    "mk",
+                    "ml",
+                    "mr",
+                    "ne",
+                    "nl",
+                    "no",
+                    "pa",
+                    "pl",
+                    "pt",
+                    "ro",
+                    "ru",
+                    "sk",
+                    "sl",
+                    "so",
+                    "sq",
+                    "sv",
+                    "sw",
+                    "ta",
+                    "te",
+                    "th",
+                    "tl",
+                    "tr",
+                    "uk",
+                    "ur",
+                    "vi",
+                    "zh-cn",
+                    "zh-tw",
+                ],
+                default=["en"],
+            )
+
+        settings["Language"] = {
+            "valid_languages": st_lan_valid_language,
         }
 
     if "PromptInjection" in st_enabled_scanners:
@@ -357,6 +432,9 @@ def get_scanner(scanner_name: str, vault: Vault, settings: Dict):
             denied_languages = settings["languages"]
 
         return Code(allowed=allowed_languages, denied=denied_languages)
+
+    if scanner_name == "Language":
+        return Language(valid_languages=settings["valid_languages"])
 
     if scanner_name == "PromptInjection":
         return PromptInjection(threshold=settings["threshold"])