+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `Regex` scanner to the prompt
+- `Language` scanners both for prompt and output
 
 ### Fixed
 -